requirements.txt
requests==2.32.3
psycopg2==2.9.9
kafka-python==2.0.2
dbt-core==1.8.2
dbt-postgres==1.8.1
dbt-snowflake==1.8.3